refactor(netflix/canary): Convert canaryScore to react
