+++ app/scripts/modules/netflix/canary/canary.less
@@ -56,7 +56,11 @@
   flex-direction: row;
 }
 
-.score-large .score {
+.score-large.score {
+  height: 48px;
+}
+
+.score-large .score.label {
   height: 48px;
   line-height: 48px;
   padding: 0 10px 0 10px;
@@ -64,6 +68,7 @@
   border-radius: 0;
   margin: 0 0 5px 10px;
   min-width: 48px;
+  display: inline-block;
 }
 
 .canary-column {
