+++ app/scripts/modules/netflix/pipeline/stage/acaTask/acaTaskStage.module.js
@@ -3,13 +3,14 @@
 let angular = require('angular');
 
 import {ACCOUNT_SERVICE} from 'core/account/account.service';
+import {CANARY_SCORE_COMPONENT} from '../canary/canaryScore.component';
 import {NAMING_SERVICE} from 'core/naming/naming.service';
 
 module.exports = angular.module('spinnaker.netflix.pipeline.stage.genericCanary', [
   require('./acaTaskStage'),
   require('./acaTaskExecutionDetails.controller'),
   require('./acaTaskStage.transformer'),
-  require('../canary/canaryScore.directive.js'),
+  CANARY_SCORE_COMPONENT,
   require('../canary/canaryStatus.directive.js'),
   ACCOUNT_SERVICE,
   NAMING_SERVICE,
