+++ app/scripts/modules/netflix/pipeline/stage/canary/canaryStage.js
@@ -6,7 +6,7 @@ let angular = require('angular');
 import {CLOUD_PROVIDER_REGISTRY} from 'core/cloudProvider/cloudProvider.registry';
 import {SERVER_GROUP_COMMAND_BUILDER_SERVICE} from 'core/serverGroup/configure/common/serverGroupCommandBuilder.service';
 import {LIST_EXTRACTOR_SERVICE} from 'core/application/listExtractor/listExtractor.service';
-import {CANARY_SCORE_CONFIG_COMPONENT} from './canaryScore.component';
+import {CANARY_SCORE_CONFIG_COMPONENT} from './canaryScores.component';
 import {NetflixSettings} from '../../../netflix.settings';
 
 module.exports = angular.module('spinnaker.netflix.pipeline.stage.canaryStage', [
